Remove clap as a dependency from the Library
We only want to have the main binary to have the clap crate imported.
This is because it's a command line arg parsing crate. Ideally we would
pass values from it to methods as needed and not pass clap's data types
functions outside of main.

This commit does just that by removing clap from lib.rs and generating
a PathBuf for the Config struct in main.rs instead.

Closes #20