Improve travis-ci with rustfmt and eslint
Currently we had no lint to check that someone had run rustfmt on the
code. This makes sure that someone runs rustfmt so they can pass the CI
checks now. We also wanted eslint to run but according to the
conversation in issue #17 this is already the case when tests are run
using the ember-cli tool.

This adds a ci-scripts folder that we can fill with more scripts we need
run during testing while not cluttering up travis with all kinds of
conditionals and things. In this case we add the rustfmt.sh script which
checks if rustfmt is already installed from the cache and if not to
install it. If it's out of date compared to upstream it'll install the
newest version of rustfmt automatically then cache it for later!

This commit also reformats lib.rs as it was not compliant with rustfmt
and was causing the CI to fail under then new changes