Remove hardcoding of Assets from rustdoc
This adds a build script to dynamically generate the assets vector used
in the creation of a new config. It does so by finding all of the assets
and creating an expression in src/assets.in. This expression looks much
like what the `Config` implementation had originally, mostly using
a name and `include_str!()`. Now though we don't have to worry about
changing the names every time as a result!

A few changes have also been made to the CI:

- Namely it tests that it can run on the example crate after having been
  run
- Makes sure rustfmt is run on the build.rs script
- Makes sure rustfmt is run on all .rs files in src only
- Dist files are generated each time to make sure it works dynamically
- Frontend tests are run first to make sure that the back will have
  files to include later on

While this still means the files are tracked and we still depend on
ember the crate can still be compiled as is until a better solution is
found. This also generates the latest assets available and checks them
into the source tree.

Closes #8
Closes #18