Inject config into generate_analysis
Previously we were calling generate_analysis in the config's
constructor. This gave generate_analysis a lot of to do; creating the
AnalysisHost and loading up the analysis data. In order to break this up a bit,
this initializes the AnalysisHost in the config and then passes the
entire config to generate_analysis. Since generate_analysis was taking
in the manifest_path, a config field, and the host, another config
field, it seemed to make sense to pass the entire struct into this
function. This also matches the high level spec that @steveklabnik
mentioned in
https://github.com/steveklabnik/rustdoc/issues/11#issuecomment-315800899.