Add error-chain to handle errors better
We were using `Box<std::error::Error>` for all of our error handling.
While not bad it makes dealing with them a bit of a pain and error-chain
handles errors very well, making it easy to create our own as well as
interface with other libraries errors. On top of that they end up
chaining well producing a stack of what caused what errors from the
point of origin all the way to the top level. This commit implements all
of the above to give us better more robust error handling.