Add identicaf for more dynamic status updates
We used to just `println!()` stuff at certain points in the code saying
this was happening etc. However, this isn't really all to pretty and it
doesn't really indicate if something is in progress. This adds the
library identicaf to show spinners as things are going on and updating
the status of things as items are completed.

Closes #37