Create missing asset sub-directories
The build process was failing when trying to include the `ember-fetch`
directory. The failure we related to the fact that we were trying to
create the `ember-fetch/fastboot-fetch-a.js` file without the
`ember-watch` director existing in `target/doc`.

Now `create_asset_file` assumes that `name` may be a path with
directories. It will first try to create the necessary directories
before trying to create the file.

Fixes #53