Remove jsonapi and replace it with serde
This changes rustdoc over completely to serde. While jsonapi was a great
crate to get started it had many features and things we weren't going to
use at all. Instead we implemented a limited subset of the spec. This
moves us over to just using serde to output data about the crate. A few
things worth noting:

- The `Crate` type was merged into the `DocData` type since it was only
  used in that type. This made the resulting serialization code much
  cleaner by removing references to krate.
- The builder method was used as a way to cleanly build up the types
  for eventual serialization and because it's readable.
- Using `&str` vs String was chosen to avoid allocations that weren't
  needed
- `HashMaps` were made with the capacity of given constants since we
  know how many items at most will exist there.

Overall this commit makes serializing data easier to work with and
more reabable then before since we don't need to use all of jsonapi's
more advanced features. The nice thing is we can still use jsonapi to
validate the output that we create!

Closes #69