Parallelize the recur function
Prior to this recur kept being called. This might have led to a blown
stack on larger projects (still unsure). However, it wasn't all too fast
to run and could easily have been parallelized. This commit does just
that by grabbing all of the analysis in parallel for everything
possible. It then reduces them into a single Vec and then extends are
already existing top level Vec into one just like before.

This was blazingly fast in release mode but debug seemed to be a bit
slow (most likely because iterators usually are in debug).