Change --manifest-path to point to Cargo.toml.
Make the `--manifest-path` option consistent with Cargo's by having it
point to the `Cargo.toml`, rather than the directory containing the
`Cargo.toml`.

If the `--manifest-path` option is not provided, the default is still to
look for a `Cargo.toml` in the current working directory. This is
different behavior to Cargo, which attempts to look for a `Cargo.toml`
by traversing up through the current working directory and its parents.
A future enhancement could fix this, but it would make sense to wait for
the outcome of #155.

Fixes #113.