Properly nest definitions
Before this commit, if you had a module structure like this:

    mod foo {
        mod bar {
            mod baz {
	    }
        }
    }

Rustdoc would not properly generate the JSON needed. With this
commit, it now does.

We introduce a work queue via a channel. The parent item is processed,
and then each child is pushed onto the queue. We then loop, taking each
item off of the queue, and pushing any children on. When the queue is
empty, we're done.