Update README.md
Our README was out of date. This updates it to have the current state
of the project reflected. It also cleans up the links so that it doesn't
pollute the reading of the text and cleans up a few other spellings,
switches '\' paths to '/' as both Windows and Unix systems can handle
'/' but both can't handle '\' which can be confusing to some users if
they just copy and paste the paths.